backend/requirements.txt
fastapi
uvicorn
networkx
google-generativeai
pydantic
python-dotenv
motor
python-multipart
requests
beautifulsoup4
